Separate script for testing utility/helper functs; bolo overlay test
- Splits test data loader utilities into a separate script for easier sharing between tests which might want these sample data
- Adds test data loader for bolometers
- Adds unittests for bolometer overlays
